pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "wheel"]

[project]
name = "useful"
description = "Miscellaneous utilities"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Python :: 3",
]
keywords = []
license = { text = "BSD 3-Clause License" }
authors = [
    { email = "diane.adjavon@icloud.com", name = "Diane Adjavon" },
]
dynamic = ["version"]
dependencies = [
  "psycopg2-binary",
  "gunpowder",
  "numpy",
  "funlib.persistence",
  "torch",
]

[project.optional-dependencies]
dev = [
  'pytest',
  'pytest-cov',
  'black',
  'mypy',
  'pdoc',
  'pre-commit'
]

[project.urls]
homepage = "https://github.com/adjavon/useful"
repository = "https://github.com/adjavon/useful"
